split firebase.ts into 2 files
