+++ src/lib/firebase/firebase-client.ts
@@ -1,11 +1,14 @@
-import { initializeApp, getApps, getApp, FirebaseApp } from "firebase/app";
+"use client";
+
+import { initializeApp, getApps, getApp, type FirebaseApp } from "firebase/app";
 import {
   initializeFirestore,
   persistentLocalCache,
   persistentMultipleTabManager,
 } from "firebase/firestore";
 
-const firebaseConfig = {
+// Only expose NEXT_PUBLIC_* variables to the client
+const clientConfig = {
   apiKey: process.env.NEXT_PUBLIC_FIREBASE_API_KEY!,
   authDomain: process.env.NEXT_PUBLIC_FIREBASE_AUTH_DOMAIN!,
   projectId: process.env.NEXT_PUBLIC_FIREBASE_PROJECT_ID!,
@@ -14,10 +17,12 @@ const firebaseConfig = {
   appId: process.env.NEXT_PUBLIC_FIREBASE_APP_ID!,
 };
 
-// Always initialize or reuse the app here
-export const app: FirebaseApp =
-  getApps().length > 0 ? getApp() : initializeApp(firebaseConfig);
+// Reuse app if already initialized
+export const app: FirebaseApp = getApps().length
+  ? getApp()
+  : initializeApp(clientConfig);
 
+// Enable persistent cache for browser clients
 export const db = initializeFirestore(app, {
   localCache: persistentLocalCache({
     tabManager: persistentMultipleTabManager(),
